requirements.txt
anthropic>=0.84.0
python-dotenv>=1.0.0
yfinance>=1.2.0
pillow>=10.0.0
